main: Move daemon to own file
